=More Haskell Copilot work

+++ extras/haskell/test/main.c
@@ -1,8 +1,11 @@
+#include <stdio.h>
+#include <stdint.h>
+
 #include "test.h"
 
-void display(float display_arg0)
+void display(uint8_t display_arg0)
 {
-    printf("%f\n", display_arg0);
+    printf("%d\n", display_arg0);
 }
 
 int main(int argc, char ** argv)
